Move markdown parser creation logic outside render()
This logic doesn't need any of the props/state in render,
so it's safe and faster to run it on file load time rather than
on each render.

If this component wasn't mounted initially, we might not
want to run this on file load time, but instead might want
to put this logic in componentDidMount, or wrap this logic in
a function called only on the first render.